# apps/git-internals-dashboard/backend/openapi.yaml
openapi: 3.0.1
info:
  title: git-internals-dashboard API
  description: >
    SLA monitor for CS-originated GitHub issues. Authless (an upstream API
    gateway validates the caller's JWT before traffic reaches this service);
    this spec declares no security scheme.
  version: 1.0.0
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /issues:
    get:
      summary: Filterable issue list.
      operationId: getIssues
      parameters:
        - $ref: '#/components/parameters/RepoParam'
        - name: priority
          in: query
          schema:
            type: string
            maxLength: 50
        - name: state
          in: query
          schema:
            type: string
            enum: [OPEN, CLOSED]
        - name: slaState
          in: query
          schema:
            $ref: '#/components/schemas/SlaState'
        - name: status
          in: query
          schema:
            type: string
            maxLength: 50
        - name: q
          in: query
          description: Issue number.
          schema:
            type: string
            pattern: '^\d{1,10}$'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
        - name: bucket
          in: query
          schema:
            type: string
            enum: [all, violated, at_risk, on_track, cs, product_side, tracked, untracked, attention]
        - name: order
          in: query
          schema:
            type: string
            enum: [budget_desc, updated_desc]
            default: updated_desc
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Issue'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /issues/{id}:
    get:
      summary: One issue plus its full status-event timeline (no actors).
      operationId: getIssue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    RepoParam:
      name: repo
      in: query
      description: "owner/name"
      schema:
        type: string
        pattern: '^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$'
  responses:
    BadRequest:
      description: validation_failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    NotFound:
      description: not_found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
    InternalServerError:
      description: internal
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
  schemas:
    ErrorPayload:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [validation_failed, not_found, sync_in_progress, sync_token_missing, internal]
            message:
              type: string
    SlaState:
      type: string
      enum: [NO_SLA, OK, AT_RISK, VIOLATED, TERMINAL]
...
    IssueSla:
      type: object
      nullable: true
      properties:
        budgetHours:
          type: number
          nullable: true
        consumedHours:
          type: number
        remainingHours:
          type: number
          nullable: true
        pctConsumed:
          type: number
          nullable: true
        slaState:
          $ref: '#/components/schemas/SlaState'
        slaRunning:
          type: boolean
    Issue:
      type: object
      properties:
        id:
          type: integer
        number:
          type: integer
        state:
          type: string
          enum: [OPEN, CLOSED]
        url:
          type: string
          nullable: true
        repo:
          type: string
        priority:
          type: string
          nullable: true
        currentStatus:
          type: string
          nullable: true
        githubCreatedAt:
          type: string
          format: date-time
          nullable: true
        githubUpdatedAt:
          type: string
          format: date-time
          nullable: true
        sla:
          $ref: '#/components/schemas/IssueSla'
    IssueEvent:
      type: object
      properties:
        id:
          type: integer
        previousStatus:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        occurredAt:
          type: string
          format: date-time
    IssueDetail:
      allOf:
        - $ref: '#/components/schemas/Issue'
        - type: object
          properties:
            events:
              type: array
              items:
                $ref: '#/components/schemas/IssueEvent'
...